write entire GPS line to disk instead of parsing
Previously the GPS data was parsed into a struct and only certain fields
were written to the log file. There's no reason to do this while the
balloon is in the air, so instead just write the entire line to disk and
then it can be parsed later.

Also, slight refactor with constants moved to a header file and all
project source files moved to the HAB directory.